:sparkles: add modal esc key to close modal
