+++ .npmrc
@@ -1,1 +1,2 @@
-registry=https://registry.npmjs.org/
+#registry=https://registry.npmjs.org/
+registry=https://registry.npm.taobao.org/
