+++ VERSION.md
@@ -1,1 +1,1 @@
-0.0.1-alpha.6
+0.0.2
