+++ components/index.tsx
@@ -1,6 +1,6 @@
-export { default as Button } from './button/button'
+export { default as Button } from './button/Button'
 
-export { default as Modal } from './modal/modal'
+export { default as Modal } from './modal/Modal'
 
 export { default as Input } from './input/input'
