+++ components/modal/style/Modal.scss
