+++ components/provider/components.tsx
@@ -2,5 +2,6 @@ export default {
   KEYCAP: 'keycap',
   ALTER: 'alter',
   BUTTON: 'btn',
-  INPUT: 'input'
+  INPUT: 'input',
+  Modal: 'modal'
 }
